fix: move terminators test to correct folder
